+++ src/pages/categoryOptionGroups/Form.spec.tsx
@@ -314,6 +314,11 @@ describe('Category option groups form tests', () => {
                 categoryOptions[1].displayName,
                 screen
             )
+            await uiActions.pickRadioField(
+                'dataDimensionType',
+                'Disaggregation',
+                screen
+            )
 
             await uiActions.submitForm(screen)