+++ src/pages/dataSets/form/DataFormContents.tsx
@@ -57,6 +57,7 @@ export const DataFormContents = React.memo(function DataFormContents({
                     filterPlaceholderPicked={i18n.t(
                         'Filter selected indicators'
                     )}
+                    maxSelections={Infinity}
                 />
             </StandardFormField>
         </SectionedFormSection>